# Infrared-Vision/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AnalyzeSceneRequestCenter = {
  /**
     * @minimum -90
     * @maximum 90
     */
  lat: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  lng: number;
};

export type AnalyzeSceneRequestViewport = {
  /** @minimum 1 */
  width?: number;
  /** @minimum 1 */
  height?: number;
};

export interface AnalyzeSceneRequest {
  /** Optional supported image data URL for the visible viewport */
  image?: string;
  /** Optional raw base64 image payload */
  imageBase64?: string;
  /** @pattern ^image/(jpeg|jpg|png|webp)$ */
  mimeType?: string;
  /**
     * @minimum -90
     * @maximum 90
     */
  lat?: number;
  /**
     * @minimum -180
     * @maximum 180
     */
  lng?: number;
  center: AnalyzeSceneRequestCenter;
  /**
     * @minimum 1
     * @maximum 22
     */
  zoom?: number;
  band?: string;
  bandMode?: string;
  composite?: string;
  filterDescription?: string;
  mapProvider?: string;
  viewport?: AnalyzeSceneRequestViewport;
}

export type AnalyzeSceneResponseResult = { [key: string]: unknown };

export interface AnalyzeSceneResponse {
  success: boolean;
  result: AnalyzeSceneResponseResult;
}

export type SafeIntegrationStatusStatus = typeof SafeIntegrationStatusStatus[keyof typeof SafeIntegrationStatusStatus];


export const SafeIntegrationStatusStatus = {
  connected: 'connected',
  configured: 'configured',
  not_configured: 'not_configured',
  error: 'error',
} as const;

export interface SafeIntegrationStatus {
  provider: string;
  connected: boolean;
  configured: boolean;
  status: SafeIntegrationStatusStatus;
  reason?: string;
}

export interface McpStatus {
  railway: SafeIntegrationStatus;
  cloudflare: SafeIntegrationStatus;
  openai: SafeIntegrationStatus;
  googleMaps: SafeIntegrationStatus;
  agentPhone: SafeIntegrationStatus;
  timestamp: string;
}

export type AgentPhoneAlertRequestChannel = typeof AgentPhoneAlertRequestChannel[keyof typeof AgentPhoneAlertRequestChannel];


export const AgentPhoneAlertRequestChannel = {
  sms: 'sms',
  voice: 'voice',
} as const;

export interface AgentPhoneAlertRequest {
  /** @pattern ^\+[1-9][0-9]{7,14}$ */
  phoneNumber: string;
  /**
     * @minLength 1
     * @maxLength 480
     */
  message: string;
  channel?: AgentPhoneAlertRequestChannel;
}

export type AgentPhoneVerificationPhoneNumbersItem = {
  number: string;
};

export interface AgentPhoneVerification {
  connected: boolean;
  authenticated: boolean;
  configured: boolean;
  provider: string;
  phoneNumbers: AgentPhoneVerificationPhoneNumbersItem[];
  reason?: string;
  timestamp: string;
}

export type AgentPhoneAlertResponseChannel = typeof AgentPhoneAlertResponseChannel[keyof typeof AgentPhoneAlertResponseChannel];


export const AgentPhoneAlertResponseChannel = {
  sms: 'sms',
  voice: 'voice',
} as const;

export interface AgentPhoneAlertResponse {
  success: boolean;
  channel: AgentPhoneAlertResponseChannel;
  messageId?: string;
  providerStatus?: string;
  reason?: string;
  timestamp: string;
}