# Infrared-Vision/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agentphone/alert:
    post:
      operationId: sendAgentPhoneAlert
      tags: [integrations]
      summary: Send a test AgentPhone alert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentPhoneAlertRequest"
      responses:
        "200":
          description: Alert result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentPhoneAlertResponse"
        "400":
          description: Invalid request
        "503":
          description: AgentPhone unavailable
components:
  schemas:
...
    AgentPhoneAlertRequest:
      type: object
      properties:
        phoneNumber:
          type: string
          pattern: "^\\+[1-9][0-9]{7,14}$"
        message: { type: string, minLength: 1, maxLength: 480 }
        channel: { type: string, enum: [sms, voice], default: sms }
      required: [phoneNumber, message]
...
    AgentPhoneAlertResponse:
      type: object
      properties:
        success: { type: boolean }
        channel: { type: string, enum: [sms, voice] }
        messageId: { type: string }
        providerStatus: { type: string }
        reason: { type: string }
        timestamp: { type: string, format: date-time }
      required: [success, channel, timestamp]