# Infrared-Vision/lib/api-zod/src/generated/types/agentPhoneAlertRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentPhoneAlertRequestChannel } from './agentPhoneAlertRequestChannel';

export interface AgentPhoneAlertRequest {
  /** @pattern ^\+[1-9][0-9]{7,14}$ */
  phoneNumber: string;
  /**
     * @minLength 1
     * @maxLength 480
     */
  message: string;
  channel?: AgentPhoneAlertRequestChannel;
}