# Infrared-Vision/lib/api-zod/src/generated/types/agentPhoneAlertResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AgentPhoneAlertResponseChannel } from './agentPhoneAlertResponseChannel';

export interface AgentPhoneAlertResponse {
  success: boolean;
  channel: AgentPhoneAlertResponseChannel;
  messageId?: string;
  providerStatus?: string;
  reason?: string;
  timestamp: Date;
}